# Improve `tspec()` performance